+++ core/index.ts
@@ -1,2 +1,4 @@
+require('./utils/polyfills');
+
 import WizerdForm from './wizerdForm';
 export default WizerdForm;