+++ core/wizerdForm.ts
@@ -304,7 +304,7 @@ export default class WizerdForm {
 	}
 
 	/**
-	 * Add a new `WizerdFromPage` by string.
+	 * Add a new `WizerdFormPage` by string.
 	 * This method can be used to create pages from AJAX Calls or other
 	 * callbacks.
 	 *